feat(home): 添加召唤师个人信息展示功能
- 在首页添加召唤师头像、昵称、等级信息展示卡片
- 集成 Naive UI 组件库实现界面元素
- 添加经验值进度条显示功能
- 实现客户端启动事件监听和数据初始化逻辑
- 添加头像动画效果样式
- 调整代码格式化配置从200到150字符宽度
- 优化设置页面按钮和标签组件布局结构

+++ src-tauri/src/lcu/types/summoner.rs
@@ -79,7 +79,7 @@ pub struct SummonerInfo {
 
     pub lv: String,
 
-    pub xp: f64,
+    pub xp: i32,
 
     pub puuid: String,
